// frontend/src/app/api/pax/model/v1ListDevicesResponse.ts
/**
 * pax/v1/service.proto
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: version not set
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { V1Device } from './v1Device';


export interface V1ListDevicesResponse { 
    devices?: Array<V1Device>;
}